Policy change:
Previously I had a policy of not modifying data from codes tables. In
this specific case, that's not useful though. With the modifications
there are only the following extra cases in the data,

  LOBBYACT , JSPACSS , CORJCOH , SPACSS , COHUC , UNK

I wrote an email to TEC (which I'm not sure will get answered) about
this. If they can clairify these remaining reports, we can add them to
this table and actively formtypecd